# Conflicts:
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mrinali/DriveToBeaconsBlue1.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mrinali/DriveToBeaconsBlue3.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mrinali/DriveToBeaconsBlueShoot.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mrinali/DriveToBeaconsRed.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mrinali/DriveToBeaconsRed1.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mrinali/DriveToBeaconsRedShoot.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mrinali/OldAutonomous/DriveToBeaconsBlue1.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mrinali/OldAutonomous/DriveToBeaconsRed1.java
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Shashank/teleop/twoControllerTeleopv6Shashank.java

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Sam/shooter/beans/ShooterMotor.java
@@ -6,7 +6,7 @@ import org.firstinspires.ftc.teamcode.Sam.shooter.util.Constants;
 
 public class ShooterMotor {
     private Constants.MOTORNAME name;
-    private double rpm = 0D;
+    private int rpm = 0;
 
     public Constants.MOTORNAME getName() {
         return name;
@@ -16,12 +16,12 @@ public class ShooterMotor {
         this.name = name;
     }
 
-    public double getRpm() {
+    public int getRpm() {
         //DbgLog.msg("rpm: "+rpm);
         return rpm;
     }
 
-    public void setRpm(double rpm) {
+    public void setRpm(int rpm) {
         this.rpm = rpm;
     }
